requirements.txt
groq==0.9.0
huggingface-hub==0.24.6
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-groq==0.1.9
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langsmith==0.1.100
llama-cloud==0.0.15
llama-index==0.11.0
llama-index-agent-openai==0.3.0
llama-index-cli==0.3.0
llama-index-core==0.11.0.post1
llama-index-embeddings-langchain==0.2.1
llama-index-embeddings-openai==0.2.2
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.0
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.0
llama-index-readers-llama-parse==0.2.0
llama-parse==0.5.0
python-dotenv==1.0.1
streamlit==1.37.1